Download ERP CORE Components

Download individual components of the ERP CORE dataset.

get_erpcore(
  component = c("n170", "mmn", "n2pc", "n400", "p3", "lrp", "ern"),
  dest_path = NULL,
  conflicts = "error",
  type = c("raw", "bids", "all")
)

get_n170(dest_path = NULL, conflicts = "error", type = c("raw", "bids", "all"))

get_mmn(dest_path = NULL, conflicts = "error", type = c("raw", "bids", "all"))

get_n2pc(dest_path = NULL, conflicts = "error", type = c("raw", "bids", "all"))

get_n400(dest_path = NULL, conflicts = "error", type = c("raw", "bids", "all"))

get_p3(dest_path = NULL, conflicts = "error", type = c("raw", "bids", "all"))

get_lrp(dest_path = NULL, conflicts = "error", type = c("raw", "bids", "all"))

get_ern(dest_path = NULL, conflicts = "error", type = c("raw", "bids", "all"))

Arguments

component
Options are:

"n170": download data for the face-specific N170 component.

"mmn": download data for the Mismatch Negativity component.

"n2pc": download data for the N2pc spatial attention component.

"n400": download data for the N400 semantic mismatch component.

"p3": download the P3 data.

"lrp": download the Lateralized Readiness Potential data.

"ern": download the Error-Related Negativity dataset.

dest_path
destination file path

conflicts
This determines what happens when a file with the same name exists at the specified destination. Can be one of the following:

"error" (the default): throw an error and abort the file transfer operation.

"skip": skip the conflicting file(s) and continue transferring the remaining files.

"overwrite": replace the existing file with the transferred copy.

type
The data is in three different formats:

"raw": download the raw data and scripts.

"bids": download the raw data in scripts in a BIDS compatible format.

"all": download the raw and processed data and scripts.

Functions

get_n170: Retrieve N170 data

get_mmn: Retrieve MMN data

get_n2pc: Retrieve N2pc data

get_n400: Retrieve N400 data

get_p3: Retrieve P3 data

get_lrp: Retrieve LRP data

get_ern: Retrieve LRP data

References

Kappenman, E.S., Farrens, J.L., Zhang, W., Stewart, A.X., & Luck, S.J. (2020). ERP CORE: An Open Resource for Human Event-Related Potential Research. PsyArxiv. https://doi.org/10.31234/osf.io/4azqm

Author

Matt Craddock matt@mattcraddock.com